% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{bcp_azure_inrdt}
\alias{bcp_azure_inrdt}
\title{bcp_azure_inrdt
copy dt from r into db}
\usage{
bcp_azure_inrdt(dt, tb, db, qy_fmt = NULL, overwrite = TRUE, id = NULL,
  id_unique = TRUE)
}
\arguments{
\item{db:}{list of srv, usr, pwd, and dbn}
}
\description{
bcp_azure_inrdt
copy dt from r into db
}

