% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{bcp_azure_query}
\alias{bcp_azure_query}
\title{bcp_azure_query
copy qy from db into fn}
\usage{
bcp_azure_query(qy, db, fn)
}
\arguments{
\item{db:}{list of srv, usr, pwd, and dbn}
}
\description{
bcp_azure_query
copy qy from db into fn
}

