% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{bcp_azure_table}
\alias{bcp_azure_table}
\title{bcp_azure_table
copy tb from db into fn
must connect to designated azsql database within azdb
because azsql server dsnt allow db change after login}
\usage{
bcp_azure_table(tb, db, fn)
}
\arguments{
\item{db:}{list of srv, usr, pwd, and dbn}
}
\description{
bcp_azure_table
copy tb from db into fn
must connect to designated azsql database within azdb
because azsql server dsnt allow db change after login
}

