% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.r
\name{calculate_dist_wlatlng}
\alias{calculate_dist_wlatlng}
\title{calculate_dist_wlatlng}
\usage{
calculate_dist_wlatlng(.lat1, .lng1, .lat2, .lng2, .measure = "mi")
}
\arguments{
\item{.measure}{- mi or km}
}
\description{
calculate dist between coordinate1<lat1, lng1> and coordinate2<lat2, lng2>
}

