% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfile.r
\name{copy_dbfile}
\alias{copy_dbfile}
\title{copy_dbfile
make a copy of dbfile into dbcopy if dbfile exist and no dbcopy
dbcopy does not have size and mode and mtime all same as dbfile}
\usage{
copy_dbfile(dbfile, dbcopy, overwrite = TRUE, copy.date = TRUE)
}
\description{
copy_dbfile
make a copy of dbfile into dbcopy if dbfile exist and no dbcopy
dbcopy does not have size and mode and mtime all same as dbfile
}

