% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfile.r
\name{create_load_dbfile}
\alias{create_load_dbfile}
\title{create_load_dbfile
create and load dbfile into memory}
\usage{
create_load_dbfile(qy, db, fn, rfrh, colname, coltype, sep = "\\t",
  header = FALSE, skip = 0L, nrows = -1L, stringsAsFactors = FALSE)
}
\description{
create_load_dbfile
create and load dbfile into memory
}

