% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfile.r
\name{create_load_dbfile_disk}
\alias{create_load_dbfile_disk}
\title{create_load_dbfile_disk
create and load dbfile into db_disk}
\usage{
create_load_dbfile_disk(qy, db, fn, rfrh, qs, db_disk)
}
\description{
create_load_dbfile_disk
create and load dbfile into db_disk
}

