% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{dtRfrhTb}
\alias{dtRfrhTb}
\title{dtRfrhTb
update dt with tb based on id on columns in vd (all none id column in tb)
nomatched dt id will be value specified in nomatch (kept intact as in dt)
nofounded tb id will be added into dt with value in tb when nofound = TRUE}
\usage{
dtRfrhTb(dt, tb, id, vd = NULL, nomatch = NULL, nofound = FALSE,
  id_unique = TRUE, in_situ = TRUE)
}
\description{
dtRfrhTb
update dt with tb based on id on columns in vd (all none id column in tb)
nomatched dt id will be value specified in nomatch (kept intact as in dt)
nofounded tb id will be added into dt with value in tb when nofound = TRUE
}

