% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{executesc}
\alias{executesc}
\title{executesc
executesc wrapper on system() to execute system commands}
\usage{
executesc(sc, ...)
}
\description{
executesc
executesc wrapper on system() to execute system commands
}
\note{
use shell() on calling and kicking off batch files
}

