% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfile.r
\name{load_dbfile}
\alias{load_dbfile}
\title{load_dbfile
load dbfile into memory w. check on file.info intact}
\usage{
load_dbfile(fn, colname, coltype, waitime = 10, maxtime = 600,
  sep = "\\t", header = FALSE, skip = 0L, nrows = -1L,
  stringsAsFactors = FALSE)
}
\description{
load_dbfile
load dbfile into memory w. check on file.info intact
}

