% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfile.r
\name{load_dbfile_sc}
\alias{load_dbfile_sc}
\title{load_dbfile_sc
load dbfile into memory w. selected columns from file}
\usage{
load_dbfile_sc(fn, colname, coltype, sc_colname, id_colname, ...)
}
\description{
load_dbfile_sc
load dbfile into memory w. selected columns from file
}

