% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbfile.r
\name{refresh_dbfile}
\alias{refresh_dbfile}
\title{refresh_dbfile
rename dbfile into dbfakv and copy dbfrfh into dbfile}
\usage{
refresh_dbfile(dbfile, dbfakv = NULL, dbfrfh = NULL)
}
\description{
refresh_dbfile
rename dbfile into dbfakv and copy dbfrfh into dbfile
}

