% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{removeWS}
\alias{removeWS}
\title{removeWS
replace whitespace "\t", "\n" and "\\s+" with single whitespace ' '.}
\usage{
removeWS(x)
}
\description{
removeWS
replace whitespace "\t", "\n" and "\\s+" with single whitespace ' '.
}

