% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{sqlite_executeqs}
\alias{sqlite_executeqs}
\title{sqlite_executeqs wrapper on sqlite3
wrapper on sqlite3 command line shell command
chooses no wrapper on the dbGetQuery function
because of dbGetQuery can not handle mutliple line query with "\n"}
\usage{
sqlite_executeqs(qs, db)
}
\description{
sqlite_executeqs wrapper on sqlite3
wrapper on sqlite3 command line shell command
chooses no wrapper on the dbGetQuery function
because of dbGetQuery can not handle mutliple line query with "\n"
}

