% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{sqlite_refreshtb}
\alias{sqlite_refreshtb}
\title{sqlite_refreshtb
refreshtb will update value when primary key exist,
and also insert new row when primary key not exist.}
\usage{
sqlite_refreshtb(dt, tb, id, db, dt_coltp = NULL, id_unique = TRUE,
  batch_size = 1000)
}
\description{
sqlite_refreshtb
refreshtb will update value when primary key exist,
and also insert new row when primary key not exist.
}

