% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{sqlite_selectidx}
\alias{sqlite_selectidx}
\title{sqlite_selectidx wrapper on dbGetQuery
select * from tb in db where id = id_value}
\usage{
sqlite_selectidx(tb, db, id, id_value)
}
\description{
sqlite_selectidx wrapper on dbGetQuery
select * from tb in db where id = id_value
}

