% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{sqlite_subsetidx}
\alias{sqlite_subsetidx}
\title{sqlite_subsetidx
fetch an subset of index table from sqlite db - inner join}
\usage{
sqlite_subsetidx(dt, tb, db, id, batch_size = 1e+05, all.dt = FALSE,
  all.tb = FALSE)
}
\arguments{
\item{dt:}{a data.frame/data.table in R with column of key}

\item{tb:}{a full reference table in sqlite with key value}
}
\value{
a data.table with value append to the R data.table
id is limit to single index - add capability to id multiple index
}
\description{
sqlite_subsetidx
fetch an subset of index table from sqlite db - inner join
}

