% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{sqlite_uploadrdt}
\alias{sqlite_uploadrdt}
\title{sqlite_uploadrdt wrapper on dbWriteTable
id and id_unique - create [unique] index}
\usage{
sqlite_uploadrdt(dt, tb, db, id = NULL, id_unique = TRUE,
  overwrite = TRUE, append = FALSE, ...)
}
\description{
sqlite_uploadrdt wrapper on dbWriteTable
id and id_unique - create [unique] index
}

