% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{sqlite_uploadrdt_dtlist}
\alias{sqlite_uploadrdt_dtlist}
\title{sqlite_uploadrdt_dtlist}
\usage{
sqlite_uploadrdt_dtlist(dtlist, tblist, db, id = NULL, id_unique = TRUE,
  index_name = NULL, overwrite = TRUE, append = FALSE, ...)
}
\description{
upload a list contains multiple dt into sqlite
a wrapper over yg::sqlite_uploadrdt()
TODO: (should) extend to list of list of dt[?]
TODO: (should) extend id into idlist [?] - [x]
}

