% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{sqodbc_createcnn}
\alias{sqodbc_createcnn}
\title{sqodbc_createcnn wrapper on odbcDriverConnect}
\usage{
sqodbc_createcnn(db)
}
\arguments{
\item{db:}{list of [dsn] srv usr pwd and dbn}
}
\description{
sqodbc_createcnn wrapper on odbcDriverConnect
}

