% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{sqodbc_executeqy}
\alias{sqodbc_executeqy}
\title{sqodbc_executeqy wrapper on sqlQuery}
\usage{
sqodbc_executeqy(qy, db, ...)
}
\arguments{
\item{db:}{list of [dsn] srv usr pwd and dbn}
}
\description{
sqodbc_executeqy wrapper on sqlQuery
}

