% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{sqodbc_existsqtb}
\alias{sqodbc_existsqtb}
\title{sqodbc_existsqtb}
\usage{
sqodbc_existsqtb(tb, db, tb_type = c("table", "view"), ...)
}
\arguments{
\item{db:}{list of [dsn] srv usr pwd and dbn}
}
\description{
sqodbc_existsqtb
}

