% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{sqodbc_refreshtb}
\alias{sqodbc_refreshtb}
\title{sqodbc_refreshtb
refreshtb mannual insert rows
refreshtb will update value when primary key exist,
and also insert new row when primary key not exist.}
\usage{
sqodbc_refreshtb(dt, tb, id, db, batch_size = 10000)
}
\description{
sqodbc_refreshtb
refreshtb mannual insert rows
refreshtb will update value when primary key exist,
and also insert new row when primary key not exist.
}

