% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{sqodbc_selectidx}
\alias{sqodbc_selectidx}
\title{sqodbc_selectidx wrapper on sqlQuery
select * from tb in db where id = id_value}
\usage{
sqodbc_selectidx(tb, db, id, id_value)
}
\description{
sqodbc_selectidx wrapper on sqlQuery
select * from tb in db where id = id_value
}

