% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{sqodbc_subsetidx}
\alias{sqodbc_subsetidx}
\title{sqodbc_subsetidx
fetch an subset of index table from sqodbc db - inner join}
\usage{
sqodbc_subsetidx(dt, tb, db, id, batch_size = 10000, all.dt = FALSE,
  all.tb = FALSE)
}
\arguments{
\item{dt:}{a data.frame/data.table in R with column of key}

\item{tb:}{a full reference table in sqodbc with key value}
}
\value{
a data.table with value append to the R data.table
id limit to single index: add capability to id multi index
sqodbc_subsetidx is a simplier verison of sqlite_subsetidx
without checking on type of each column when define dt_rtn
}
\description{
sqodbc_subsetidx
fetch an subset of index table from sqodbc db - inner join
}

