% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{sqodbc_updatestb}
\alias{sqodbc_updatestb}
\title{sqodbc_updatestb wrapper on sqlUpdate
sqlUpdate will update value when primary key exist,
but no insert or append when primary key not exist.}
\usage{
sqodbc_updatestb(dt, tb, id, db, ...)
}
\description{
sqodbc_updatestb wrapper on sqlUpdate
sqlUpdate will update value when primary key exist,
but no insert or append when primary key not exist.
}

