% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite.r
\name{sqodbc_uploadrdt}
\alias{sqodbc_uploadrdt}
\title{sqodbc_uploadrdt wrapper on sqlSave
make sqodbc_uploadrdt wrapper on bcp_azure_inrdt?}
\usage{
sqodbc_uploadrdt(dt, tb, db, id = NULL, id_unique = TRUE,
  rownames = FALSE, colnames = FALSE, safer = TRUE, fast = TRUE,
  overwrite = TRUE, ...)
}
\description{
sqodbc_uploadrdt wrapper on sqlSave
make sqodbc_uploadrdt wrapper on bcp_azure_inrdt?
}

